# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MobileWallet backend",
    "description": "\nBackend **MobileWallet** — un rassemblement d'**agrégateurs** de paiement Mobile Money.\n\nChaque agrégateur (DigiKUNTZ aujourd'hui, d'autres demain) est un module exposant\ndeux capacités :\n\n- **navigateur IA** (`mode=browser`) : un agent Playwright + LLM pilote le checkout,\n  exécute le paiement et **déduit** le « curl replay » (requête `/charge` + `/verify`),\n  persisté comme *template* réutilisable.\n- **curl replay** (`mode=replay`) : rejoue le paiement sans navigateur via le template stocké.\n- **auto** (`mode=auto`, défaut) : tente le replay d'abord, puis bascule sur le navigateur\n  si le replay est non concluant (`fallback_browser`).\n\nToute tentative est auditée (table `transactions`) quand Supabase est configuré.\n",
    "version": "0.2"
  },
  "paths": {
...
    "/pay": {
      "post": {
        "tags": [
          "payments"
        ],
        "summary": "Exécuter un paiement",
        "description": "Exécute un paiement via l'agrégateur et le mode demandés.\n\nAuthentifié par clé API (`Authorization: Bearer <clé>`). La transaction est\nrattachée à l'app de la clé (isolation) ; `end_user_ref` est conservé tel quel.\n\n- **auto** (défaut) : replay d'abord ; bascule navigateur si non concluant\n  (selon `fallback_browser`).\n- **browser** : flux IA complet ; déduit et persiste le template curl.\n- **replay** : rejoue via le template stocké (409 si absent).",
        "operationId": "pay_pay_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PayResponse"
                }
              }
            }
          },
          "404": {
            "description": "Agrégateur inconnu."
          },
          "400": {
            "description": "Mode invalide."
          },
          "422": {
            "description": "Réseau non supporté (renvoie la liste exacte attendue)."
          },
          "409": {
            "description": "Mode replay sans template (lancer mode=browser d'abord)."
          },
          "502": {
            "description": "Replay échoué et fallback navigateur désactivé."
          },
          "503": {
            "description": "Service de paiement amont temporairement indisponible (code=network_unavailable). Réessayer plus tard."
          }
        }
      }
    },
...
    "/status/{transaction_ref}": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "Statut d'une transaction",
        "description": "Récupère une transaction de votre app par sa référence.",
        "operationId": "transaction_status_status__transaction_ref__get",
        "parameters": [
          {
            "name": "transaction_ref",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Transaction Ref"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Aucune transaction pour cette référence."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PayRequest": {
        "properties": {
          "amount": {
            "type": "integer",
            "title": "Amount",
            "description": "Montant en XAF.",
            "examples": [
              25
            ]
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "description": "Numéro Mobile Money (sans +237).",
            "examples": [
              "696080087"
            ]
          },
          "network": {
            "type": "string",
            "title": "Network",
            "description": "Réseau : Orangemoney ou MTN.",
            "examples": [
              "Orangemoney"
            ]
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "Email du payeur.",
            "examples": [
              "client@example.com"
            ]
          },
          "sender_name": {
            "type": "string",
            "title": "Sender Name",
            "description": "Nom affiché de l'émetteur.",
            "default": "Rauvalia"
          },
          "callback_url": {
            "type": "string",
            "title": "Callback Url",
            "description": "URL de callback (défaut: celle de l'agrégateur).",
            "default": ""
          },
          "aggregator": {
            "type": "string",
            "title": "Aggregator",
            "description": "Nom de l'agrégateur (cf. GET /aggregators).",
            "default": "digikuntz"
          },
          "mode": {
            "type": "string",
            "title": "Mode",
            "description": "auto | browser | replay.",
            "default": "auto",
            "examples": [
              "auto"
            ]
          },
          "fallback_browser": {
            "type": "boolean",
            "title": "Fallback Browser",
            "description": "En mode auto : basculer sur le navigateur si le replay est non concluant.",
            "default": true
          },
          "end_user_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End User Ref",
            "description": "Identifiant opaque de votre utilisateur final (ex. votre user_id). Stocké tel quel pour vos rapprochements ; facultatif."
          }
        },
        "type": "object",
        "required": [
          "amount",
          "phone",
          "network",
          "email"
        ],
        "title": "PayRequest",
        "examples": [
          {
            "aggregator": "digikuntz",
            "amount": 25,
            "email": "client@example.com",
            "fallback_browser": true,
            "mode": "auto",
            "network": "Orangemoney",
            "phone": "696080087"
          }
        ]
      },
      "PayResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Vrai si le paiement a abouti."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "successful | ussd_sent | failed | cancelled | pending.",
            "default": ""
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Message clair pour l'utilisateur final.",
            "default": ""
          },
          "transaction_id": {
            "type": "string",
            "title": "Transaction Id",
            "description": "Référence transaction de l'agrégateur.",
            "default": ""
          },
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "Code HTTP de la réponse (200 si OK).",
            "default": 200
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "PayResponse",
        "description": "Vue CLIENT (dev qui intègre l'API) — strictement l'essentiel métier.\n\nTout le détail technique (charge_url, plaintext, curl, trace, tokens…) est\nréservé à l'admin (voir PayResponseDebug + ?debug=true). Ne JAMAIS ajouter de\nchamp interne ici."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Clé API d'app (paiements) : Authorization: Bearer <sk_…>."
      },
      "DevJWT": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT du compte developer (gestion des projets) : Authorization: Bearer <jwt>."
      },
      "AdminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Key",
        "description": "Clé admin pour la supervision des developers/apps/clés."
      }
    }
  },
...
}